Updated behaviour tree for team AI that allows the AIs to transition between manoeuvers based on the preconditions etc. However, there seems to be some memory issue arising from these transitions as more or less random errors pop up. This is a safety copy to be able to rewind to this version if too much gets messed up in debugging without success.

+++ SquadAI/RushBaseAttack.cpp
@@ -218,7 +218,7 @@ void RushBaseAttack::DetermineAssemblyPoint(void)
 
 	do
 	{
-		float randomAngle = XMConvertToRadians(rand() % 180);
+		float randomAngle = XMConvertToRadians(static_cast<float>(rand() % 180));
 		
 		float sine = sin(randomAngle);
 		float cosine = cos(randomAngle);
